+++ crates/core/src/addon.rs
@@ -1,5 +1,5 @@
 use crate::{
-    error::{ParseError, RepositoryError},
+    error::ParseError,
     repository::{
         ReleaseChannel, RemotePackage, RepositoryIdentifiers, RepositoryKind, RepositoryMetadata,
         RepositoryPackage,
@@ -124,12 +124,6 @@ pub struct Addon {
     #[cfg(feature = "gui")]
     pub details_btn_state: iced_native::button::State,
     #[cfg(feature = "gui")]
-    pub remote_btn_state: iced_native::button::State,
-    #[cfg(feature = "gui")]
-    pub local_btn_state: iced_native::button::State,
-    #[cfg(feature = "gui")]
-    pub full_changelog_btn_state: iced_native::button::State,
-    #[cfg(feature = "gui")]
     pub update_btn_state: iced_native::button::State,
     #[cfg(feature = "gui")]
     pub force_btn_state: iced_native::button::State,
@@ -143,6 +137,8 @@ pub struct Addon {
     pub website_btn_state: iced_native::button::State,
     #[cfg(feature = "gui")]
     pub pick_release_channel_state: iced_native::pick_list::State<ReleaseChannel>,
+    #[cfg(feature = "gui")]
+    pub changelog_btn_state: iced_native::button::State,
 }
 
 impl Addon {
@@ -157,12 +153,6 @@ impl Addon {
             #[cfg(feature = "gui")]
             details_btn_state: Default::default(),
             #[cfg(feature = "gui")]
-            remote_btn_state: Default::default(),
-            #[cfg(feature = "gui")]
-            local_btn_state: Default::default(),
-            #[cfg(feature = "gui")]
-            full_changelog_btn_state: Default::default(),
-            #[cfg(feature = "gui")]
             update_btn_state: Default::default(),
             #[cfg(feature = "gui")]
             force_btn_state: Default::default(),
@@ -176,6 +166,8 @@ impl Addon {
             website_btn_state: Default::default(),
             #[cfg(feature = "gui")]
             pick_release_channel_state: Default::default(),
+            #[cfg(feature = "gui")]
+            changelog_btn_state: Default::default(),
         }
     }
 
@@ -333,6 +325,13 @@ impl Addon {
     /// Returns the website url of the addon.
     pub fn website_url(&self) -> Option<&str> {
         self.metadata().map(|m| m.website_url.as_deref()).flatten()
+    }
+
+    /// Returns the changelog url of the addon.
+    pub fn changelog_url(&self) -> Option<&str> {
+        self.metadata()
+            .map(|m| m.changelog_url.as_deref())
+            .flatten()
     }
 
     /// Returns the curse id of the addon, if applicable.
@@ -496,19 +495,6 @@ impl Addon {
                 alpha_package
             }
         }
-    }
-
-    pub async fn get_changelog(
-        &self,
-        is_remote: bool,
-    ) -> Result<(String, String), RepositoryError> {
-        if let Some(repository) = self.repository() {
-            return repository
-                .get_changelog(self.release_channel, is_remote)
-                .await;
-        }
-
-        Err(RepositoryError::AddonNoRepository)
     }
 }