+++ crates/core/src/utility.rs
@@ -25,21 +25,6 @@ pub(crate) fn strip_non_digits(string: &str) -> Option<String> {
     Some(stripped)
 }
 
-pub(crate) fn truncate(s: &str, max_chars: usize) -> &str {
-    match s.char_indices().nth(max_chars) {
-        None => s,
-        Some((idx, _)) => &s[..idx],
-    }
-}
-
-pub(crate) fn regex_html_tags_to_newline() -> Regex {
-    regex::Regex::new(r"<br ?/?>|#.\s").unwrap()
-}
-
-pub(crate) fn regex_html_tags_to_space() -> Regex {
-    regex::Regex::new(r"<[^>]*>|&#?\w+;|[gl]t;").unwrap()
-}
-
 #[derive(Debug, Deserialize, Clone)]
 pub struct Release {
     pub tag_name: String,