+++ src/gui/update.rs
@@ -1,9 +1,8 @@
 use {
     super::{
-        AddonVersionKey, Ajour, BackupFolderKind, CatalogCategory, CatalogColumnKey, CatalogRow,
-        CatalogSource, Changelog, ChangelogPayload, ColumnKey, DirectoryType, DownloadReason,
-        ExpandType, InstallAddon, InstallKind, InstallStatus, Interaction, Message, Mode,
-        SelfUpdateStatus, SortDirection, State,
+        Ajour, BackupFolderKind, CatalogCategory, CatalogColumnKey, CatalogRow, CatalogSource,
+        ColumnKey, DirectoryType, DownloadReason, ExpandType, InstallAddon, InstallKind,
+        InstallStatus, Interaction, Message, Mode, SelfUpdateStatus, SortDirection, State,
     },
     crate::{log_error, Result},
     ajour_core::{
@@ -271,47 +270,6 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                         ajour.expanded_type = expand_type.clone();
                     }
                 }
-                ExpandType::Changelog(changelog) => match changelog {
-                    // We request changelog.
-                    Changelog::Request(addon, key) => {
-                        log::debug!(
-                            "Interaction::Expand(Changelog::Request({:?}))",
-                            &addon.primary_folder_id
-                        );
-
-                        // Check if the current expanded_type is showing changelog, and is the same
-                        // addon. If this is the case, we close the details.
-
-                        if let ExpandType::Changelog(Changelog::Some(a, _, k)) =
-                            &ajour.expanded_type
-                        {
-                            if addon.primary_folder_id == a.primary_folder_id && key == k {
-                                ajour.expanded_type = ExpandType::None;
-                                return Ok(Command::none());
-                            }
-                        }
-
-                        ajour.expanded_type =
-                            ExpandType::Changelog(Changelog::Loading(addon.clone(), *key));
-                        return Ok(Command::perform(
-                            perform_fetch_changelog(addon.clone(), *key),
-                            Message::FetchedChangelog,
-                        ));
-                    }
-                    Changelog::Loading(a, _) => {
-                        log::debug!(
-                            "Interaction::Expand(Changelog::Loading({:?}))",
-                            &a.primary_folder_id
-                        );
-                        ajour.expanded_type = ExpandType::Changelog(changelog.clone());
-                    }
-                    Changelog::Some(a, _, _) => {
-                        log::debug!(
-                            "Interaction::Expand(Changelog::Some({:?}))",
-                            &a.primary_folder_id
-                        );
-                    }
-                },
                 ExpandType::None => {
                     log::debug!("Interaction::Expand(ExpandType::None)");
                 }
@@ -1371,21 +1329,6 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                 }
             }
         }
-        Message::FetchedChangelog((addon, key, result)) => {
-            log::debug!("Message::FetchedChangelog(error: {})", &result.is_err());
-            match result.context("Failed to fetch changelog") {
-                Ok((changelog, url)) => {
-                    let payload = ChangelogPayload { changelog, url };
-                    let changelog = Changelog::Some(addon, payload, key);
-                    ajour.expanded_type = ExpandType::Changelog(changelog);
-                }
-                Err(error) => {
-                    log_error(&error);
-                    ajour.error = Some(error);
-                    ajour.expanded_type = ExpandType::None;
-                }
-            }
-        }
         Message::Interaction(Interaction::UpdateAjour) => {
             log::debug!("Interaction::UpdateAjour");
 
@@ -1565,21 +1508,6 @@ async fn perform_read_addon_directory(
         flavor,
         read_addon_directory(addon_cache, fingerprint_cache, root_dir, flavor).await,
     )
-}
-
-async fn perform_fetch_changelog(
-    addon: Addon,
-    key: AddonVersionKey,
-) -> (
-    Addon,
-    AddonVersionKey,
-    Result<(String, String), RepositoryError>,
-) {
-    let is_remote = key == AddonVersionKey::Remote;
-
-    let result = addon.get_changelog(is_remote).await;
-
-    (addon, key, result)
 }
 
 /// Downloads the newest version of the addon.